+++ src/fileHandlers/index.ts
@@ -4,4 +4,5 @@ export * from './diff';
 export * from './rename';
 export * from './create';
 export * from './compareFolders';
+export * from './syncPreview';
 export { handleCtxFromUri, allHandleCtxFromUri, FileHandlerContext } from './createFileHandler';